+++ bin/ci/validate-build-bundles.sh
@@ -1,0 +1,7 @@
+#! /bin/bash
+
+changes=$(git status -s | grep dist -c)
+expected=12
+exitCode=$(expr $changes - $expected)
+echo $exitCode
+exit $exitCode